chore(conversations): split channel tasks
