+++ products/conversations/backend/api/tests/test_email_endpoints.py
@@ -1860,14 +1860,14 @@ class TestSendEmailReplyMultiConfig(BaseTest):
         return comment, outbox
 
     def _run_reply(self, ticket: Ticket, content: str = "Reply from agent") -> tuple[Comment, EmailOutboxMessage]:
-        from products.conversations.backend.tasks import send_email_reply
+        from products.conversations.backend.tasks.email import send_email_reply
 
         comment, outbox = self._create_outbox(ticket, content=content)
         send_email_reply(str(outbox.id))
         outbox.refresh_from_db()
         return comment, outbox
 
-    @patch("products.conversations.backend.tasks.send_mime")
+    @patch("products.conversations.backend.tasks.email.send_mime")
     def test_send_email_reply_uses_ticket_config(self, mock_send_mime: MagicMock):
         config1 = self._create_config("support@example.com", "aaa111")
         self._create_config("billing@example.com", "bbb222")
@@ -1908,9 +1908,9 @@ class TestSendEmailReplyMultiConfig(BaseTest):
             ("no_customer_email", "no customer email"),
         ]
     )
-    @patch("products.conversations.backend.tasks.send_mime")
+    @patch("products.conversations.backend.tasks.email.send_mime")
     def test_undeliverable_reply_fails_visibly(self, name: str, expected_error: str, mock_send_mime: MagicMock):
-        from products.conversations.backend.tasks import send_email_reply
+        from products.conversations.backend.tasks.email import send_email_reply
 
         config = self._create_config("support@example.com", "aaa111")
         if name == "email_disabled":
@@ -1954,7 +1954,7 @@ class TestSendEmailReplyMultiConfig(BaseTest):
             ("domain_not_registered", MailgunDomainNotRegistered("gone from mailgun"), True),
         ]
     )
-    @patch("products.conversations.backend.tasks.send_mime")
+    @patch("products.conversations.backend.tasks.email.send_mime")
     def test_send_email_reply_terminal_errors_mark_failed(
         self, _name: str, error: Exception, flips_domain_verified: bool, mock_send_mime: MagicMock
     ):
@@ -1972,7 +1972,7 @@ class TestSendEmailReplyMultiConfig(BaseTest):
         config.refresh_from_db()
         assert config.domain_verified is (not flips_domain_verified)
 
-    @patch("products.conversations.backend.tasks.send_mime")
+    @patch("products.conversations.backend.tasks.email.send_mime")
     def test_send_email_reply_delivers_to_team_member_ticket(self, mock_send_mime: MagicMock):
         """An in-app agent reply on a ticket opened by a team member (e.g. dogfooding
         the support inbox) must still be delivered to them."""
@@ -1987,11 +1987,11 @@ class TestSendEmailReplyMultiConfig(BaseTest):
         assert mock_send_mime.call_args[1]["recipients"] == [self.user.email]
         assert outbox.status == EmailOutboxMessage.Status.SENT
 
-    @patch("products.conversations.backend.tasks.send_mime")
+    @patch("products.conversations.backend.tasks.email.send_mime")
     def test_send_email_reply_skips_comment_from_inbound_email(self, mock_send_mime: MagicMock):
         """Last-mile echo guard: an outbox row pointing at a comment that itself arrived
         via inbound email must never be sent, even if a regression enqueues one."""
-        from products.conversations.backend.tasks import send_email_reply
+        from products.conversations.backend.tasks.email import send_email_reply
 
         config = self._create_config("support@example.com", "aaa111")
         ticket = self._create_ticket(config)
@@ -2019,7 +2019,7 @@ class TestSendEmailReplyMultiConfig(BaseTest):
         assert outbox.status == EmailOutboxMessage.Status.FAILED_PERMANENT
         assert outbox.last_error == "comment originated from inbound email"
 
-    @patch("products.conversations.backend.tasks.send_mime")
+    @patch("products.conversations.backend.tasks.email.send_mime")
     def test_send_email_reply_transient_error_schedules_retry(self, mock_send_mime: MagicMock):
         """Transient errors must NOT be dropped or raised — the row stays pending with a
         backed-off next_attempt_at so the sweeper re-drives it. This is what survives a
@@ -2038,7 +2038,7 @@ class TestSendEmailReplyMultiConfig(BaseTest):
         assert outbox.next_attempt_at > before
         assert outbox.locked_until is None
 
-    @patch("products.conversations.backend.tasks.send_mime")
+    @patch("products.conversations.backend.tasks.email.send_mime")
     def test_send_email_reply_reuses_message_id_across_attempts(self, mock_send_mime: MagicMock):
         """A retried send must reuse the same Message-ID so threading/dedup stay stable."""
         config = self._create_config("support@example.com", "aaa111")
@@ -2050,7 +2050,7 @@ class TestSendEmailReplyMultiConfig(BaseTest):
         original_message_id = outbox.message_id
 
         # Make it due again and let the next attempt succeed.
-        from products.conversations.backend.tasks import send_email_reply
+        from products.conversations.backend.tasks.email import send_email_reply
 
         EmailOutboxMessage.objects.filter(id=outbox.id).update(next_attempt_at=timezone.now(), locked_until=None)
         mock_send_mime.side_effect = None
@@ -2065,9 +2065,9 @@ class TestSendEmailReplyMultiConfig(BaseTest):
         assert original_message_id.encode() in first_mime
         assert original_message_id.encode() in second_mime
 
-    @patch("products.conversations.backend.tasks.send_mime")
+    @patch("products.conversations.backend.tasks.email.send_mime")
     def test_send_email_reply_idempotent_when_already_sent(self, mock_send_mime: MagicMock):
-        from products.conversations.backend.tasks import send_email_reply
+        from products.conversations.backend.tasks.email import send_email_reply
 
         config = self._create_config("support@example.com", "aaa111")
         ticket = self._create_ticket(config)
@@ -2078,7 +2078,7 @@ class TestSendEmailReplyMultiConfig(BaseTest):
 
         mock_send_mime.assert_not_called()
 
-    @patch("products.conversations.backend.tasks.send_mime")
+    @patch("products.conversations.backend.tasks.email.send_mime")
     def test_send_email_reply_marks_failed_when_no_config(self, mock_send_mime: MagicMock):
         ticket = self._create_ticket(None)
 
@@ -2115,11 +2115,11 @@ class TestSendEmailReplyMultiConfig(BaseTest):
             ("expired_row_is_given_up", "expired", False, EmailOutboxMessage.Status.FAILED_PERMANENT),
         ]
     )
-    @patch("products.conversations.backend.tasks.send_mime")
+    @patch("products.conversations.backend.tasks.email.send_mime")
     def test_flush_pending_email_replies(
         self, _name: str, scenario: str, expect_send: bool, expected_status: str, mock_send_mime: MagicMock
     ):
-        from products.conversations.backend.tasks import EMAIL_OUTBOX_MAX_AGE, flush_pending_email_replies
+        from products.conversations.backend.tasks.email import EMAIL_OUTBOX_MAX_AGE, flush_pending_email_replies
 
         config = self._create_config("support@example.com", "aaa111")
         ticket = self._create_ticket(config)
